Added User Metrics
User metrics is a way for postly to measure what content the user likes the most. The goal is to take this data and use it to recommend likeable content - new posts will be processed and ranked as they come in and if they fit your feed it will automatically append! all you have to do is curate what you like by engaging.